backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
ultralytics
faiss-cpu
numpy
python-multipart
pyyaml
Pillow
